Extract keyboard handlers to ExpandableMenu component
